requirements.txt
django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
gunicorn>=21.2.0
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9
whitenoise>=6.6.0
requests
djangorestframework-simplejwt
Pillow